+++ Sources/Satin/Compute/BufferComputeSystem.swift
@@ -118,7 +118,7 @@ open class BufferComputeSystem: ComputeSystem {
         guard count > 0,
               bufferMap.count > 0,
               let computeEncoder = frameCommand.commandBuffer.makeComputeCommandEncoder(),
-              let argumentTable = Metal4ComputeArgumentTable(device: device)
+              let argumentTable = Metal4ComputeArgumentTable(device: device, resourceHandler: frameCommand.useResource)
         else { return false }
 
         computeEncoder.label = label