+++ Sources/Satin/Compute/TextureComputeSystem.swift
@@ -180,7 +180,7 @@ open class TextureComputeSystem: ComputeSystem {
         guard (_reset && resetPipeline != nil) || updatePipeline != nil else { return true }
         guard textures.count > 0,
               let computeEncoder = frameCommand.commandBuffer.makeComputeCommandEncoder(),
-              let argumentTable = Metal4ComputeArgumentTable(device: device)
+              let argumentTable = Metal4ComputeArgumentTable(device: device, resourceHandler: frameCommand.useResource)
         else { return false }
 
         computeEncoder.label = label